use the same index test for del.ts but with subtractions
