+++ test/functional/indexes.ts
@@ -2,10 +2,8 @@ import * as test from "tape";
 import {
   IImage, IIndex, IQueryBody,
 } from "../../fotos/types";
-import { createIndexAdd } from "./createIndexAdjustment";
+import { createIndexChangeTable, MODES } from "./createIndexChangeTable";
 import formatError from "./formatError";
-import { getIncorrectIndexUpdates } from "./getIncorrectIndexUpdates";
-import { getItemsInImages } from "./getItemsInImages";
 
 export default function indexesTests(setupData, api) {
 
@@ -35,43 +33,8 @@ export default function indexesTests(setupData, api) {
       .catch(formatError);
   });
 
-  function createIndexChangeTable(images: IImage[], existingIndex: IIndex, actualIndex: IIndex) {
-    const people = Object.keys(actualIndex.people).map((p: string) => {
-      const imagesWithPerson = images.filter((img) => img.people && img.people.includes(p));
-      const existing: number = isNaN(existingIndex.people[p]) ? 0 : existingIndex.people[p];
-      return {
-        actual: actualIndex.people[p],
-        existing,
-        expected: existing + imagesWithPerson.length,
-        id: p,
-        images: imagesWithPerson.map((img) => img.img_key),
-      };
-    });
-    const tags = Object.keys(actualIndex.tags).map((t: string) => {
-      const imagesWithTag = images.filter((img) => img.tags && img.tags.includes(t));
-      const existing: number = isNaN(existingIndex.tags[t]) ? 0 : existingIndex.tags[t];
-      return {
-        actual: actualIndex.tags[t],
-        existing,
-        expected: existing + imagesWithTag.length,
-        id: t,
-        images: imagesWithTag.map((img) => img.img_key),
-      };
-    });
-
-    const valid = {
-      people: people.filter((p) => p.actual !== p.expected),
-      tags: tags.filter((t) => t.actual !== t.expected),
-    };
-    return {
-      people,
-      tags,
-      valid,
-    };
-  }
-
   test("get indexes, should have at least tags and people of test images", (t) => {
-
+    // need to replace this with baseline index that we grab before tests start
     const existingIndexes = {
       people: {},
       tags: {},
@@ -83,7 +46,7 @@ export default function indexesTests(setupData, api) {
       fn: api.get,
     };
     const retryableTestThen = (responseBody: IIndex) => {
-      const changes = createIndexChangeTable(testImages, existingIndexes, responseBody);
+      const changes = createIndexChangeTable(MODES.ADD, testImages, existingIndexes, responseBody);
       if (changes.valid.people.length + changes.valid.tags.length > 0) {
         if (retryCount < retryStrategy.length) {
           setTimeout(() => {
